+++ examples/standalone/todomvc-solid/src/livestore/store.tsx
@@ -1,11 +1,11 @@
-import { makeAdapter } from '@livestore/adapter-web'
+import { makePersistedAdapter } from '@livestore/adapter-web'
 import LiveStoreSharedWorker from '@livestore/adapter-web/shared-worker?sharedworker'
 import { getStore } from '@livestore/solid'
 
 import LiveStoreWorker from './livestore.worker?worker'
 import { schema } from './schema.js'
 
-const adapterFactory = makeAdapter({
+const adapterFactory = makePersistedAdapter({
   storage: { type: 'opfs' },
   worker: LiveStoreWorker,
   sharedWorker: LiveStoreSharedWorker,